Restructure how user/{cct,simple,indexable} works
