+++ src/system/colors/palettes.h
@@ -1,6 +1,7 @@
 #ifndef PALETTES_H
 #define PALETTES_H
 
+// this file is active only if LMBD_LAMP_TYPE=indexable
 #ifdef LMBD_LAMP_TYPE__INDEXABLE
 
 #include <cstdint>
