+++ src/system/colors/wipes.h
@@ -1,11 +1,11 @@
 #ifndef ANIMATIONS_WIPES_H
 #define ANIMATIONS_WIPES_H
 
+// this file is active only if LMBD_LAMP_TYPE=indexable
 #ifdef LMBD_LAMP_TYPE__INDEXABLE
 
+#include "src/system/colors/colors.h"
 #include "src/system/utils/strip.h"
-
-#include "colors.h"
 
 namespace animations {
