+++ src/user/simple_functions.cpp
@@ -55,6 +55,8 @@ bool button_hold_usermode(const uint8_t, const bool, const uint32_t) {
 
 void loop() {}
 
+bool should_spawn_thread() { return false; }
+
 void user_thread() {}
 
 }  // namespace user
